Reorganize /include/ to add a /sys/, this way we don't need to muck with #includes in existing Solaris style source to get it to find the right stuff.
git-svn-id: https://outreach.scidac.gov/svn/spl/trunk@18 7e1ea52c-4ff2-0310-8f11-9dd32ca42a1c

+++ lib/list.c
@@ -776,7 +776,7 @@ list_alloc_aux (int size, void *pfreelist)
     void **plast;
 
     assert(sizeof(char) == 1);
-    assert(size >= sizeof(void *));
+    assert(size >= (int)sizeof(void *));
     assert(pfreelist != NULL);
     assert(LIST_ALLOC > 0);
     list_mutex_lock(&list_free_lock);
